Provide default field setter implementation as a separate class
